Basic Command support
- Split out NGSI-v1, v2 and LD handlers into separate files.
- Add support for PATCH incoming commands.
- Amend tests ( replace `/v2/op/update/`  with NGSI-LD PATCH)
- Amend test expectations to NGSI-LD format
