+++ lib/services/northBound/contextServer.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Telefonica Investigación y Desarrollo, S.A.U
+ * Copyright 2020 Telefonica Investigación y Desarrollo, S.A.U
  *
  * This file is part of fiware-iotagent-lib
  *
@@ -21,1308 +21,19 @@
  * please contact with::daniel.moranjimenez@telefonica.com
  *
  * Modified by: Daniel Calvo - ATOS Research & Innovation
+ * Modified by: Jason Fox - FIWARE Foundation
  */
 'use strict';
 
-var async = require('async'),
-    apply = async.apply,
-    logger = require('logops'),
-    constants = require('../../constants'),
-    errors = require('../../errors'),
-    ngsi = require('../ngsi/ngsiService'),
-    intoTrans = require('../common/domain').intoTrans,
-    deviceService = require('../devices/deviceService'),
-    commands = require('../commands/commandService'),
-    middlewares = require('../common/genericMiddleware'),
-    _ = require('underscore'),
+var intoTrans = require('../common/domain').intoTrans,
     config = require('../../commonConfig'),
     context = {
         op: 'IoTAgentNGSI.ContextServer'
     },
-    updateContextTemplateNgsi1 = require('../../templates/updateContextNgsi1.json'),
-    updateContextTemplateNgsi2 = require('../../templates/updateContextNgsi2.json'),
-    updateContextTemplateNgsiLD = require('../../templates/updateContextNgsiLD.json'),
-    queryContextTemplate = require('../../templates/queryContext.json'),
-    notificationTemplateNgsi1 = require('../../templates/notificationTemplateNgsi1.json'),
-    notificationTemplateNgsi2 = require('../../templates/notificationTemplateNgsi2.json'),
-    notificationTemplateNgsiLD = require('../../templates/notificationTemplateNgsiLD.json'),
-    notificationMiddlewares = [],
-    updateHandler,
-    commandHandler,
-    queryHandler,
-    notificationHandler;
-
-/**
- * Create the response for an UpdateContext operation, based on the results of the individual updates. The signature
- * retains the results object for homogeinity with the createQuery* version.
- *
- * @param {Object} req              Request that was handled in first place.
- * @param {Object} res              Response that will be sent.
- * @param {Object} results          Ignored for this function. TODO: to be removed in later versions.
- * @return {{contextResponses: Array}}
- */
-function createUpdateResponse(req, res, results) {
-    var result = {
-        contextResponses: []
-    };
-
-    for (var i = 0; i < req.body.contextElements.length; i++) {
-        var contextResponse = {
-            contextElement: {
-                attributes: req.body.contextElements[i].attributes,
-                id: req.body.contextElements[i].id,
-                isPattern: false,
-                type: req.body.contextElements[i].type
-            },
-            statusCode: {
-                code: 200,
-                reasonPhrase: 'OK'
-            }
-        };
-
-        for (var j = 0; j < contextResponse.contextElement.attributes.length; j++) {
-            contextResponse.contextElement.attributes[i].value = '';
-        }
-
-        result.contextResponses.push(contextResponse);
-    }
-
-    logger.debug(context, 'Generated update response: %j', result);
-
-    return result;
-}
-
-/**
- * Create the response for a queryContext operation based on the individual results gathered from the query handlers.
- * The returned response is in the NGSI Response format.
- *
- * @param {Object} req              Request that was handled in first place.
- * @param {Object} res              Response that will be sent.
- * @param {Object} results          Individual Context Element results from the query handlers.
- * @return {{contextResponses: Array}}
- */
-function createQueryResponse(req, res, results) {
-    var result = {
-        contextResponses: []
-    };
-
-    for (var i = 0; i < results.length; i++) {
-        var contextResponse = {
-            contextElement: results[i],
-            statusCode: {
-                code: 200,
-                reasonPhrase: 'OK'
-            }
-        };
-
-        contextResponse.contextElement.isPattern = false;
-
-        result.contextResponses.push(contextResponse);
-    }
-
-    logger.debug(context, 'Generated query response: %j', result);
-
-    return result;
-}
-
-/**
- * Retrieve the Device that corresponds to a Context Update, and execute the update side effects
- * if there were any (e.g.: creation of attributes related to comands).
- *
- * @param {String} device           Object that contains all the information about the device.
- * @param {String} id               Entity ID of the device to find.
- * @param {String} type             Type of the device to find.
- * @param {String} service          Service of the device.
- * @param {String} subservice       Subservice of the device.
- * @param {Array}  attributes       List of attributes to update with their types and values.
- */
-function executeUpdateSideEffects(device, id, type, service, subservice, attributes, callback) {
-    var sideEffects = [];
-
-    if (device.commands) {
-        for (var i = 0; i < device.commands.length; i++) {
-            for (var j = 0; j < attributes.length; j++) {
-                if (device.commands[i].name === attributes[j].name) {
-                    var newAttributes = [
-                        {
-                            name: device.commands[i].name + '_status',
-                            type: constants.COMMAND_STATUS,
-                            value: 'PENDING'
-                        }
-                    ];
-
-                    sideEffects.push(
-                        apply(ngsi.update,
-                            device.name,
-                            device.resource,
-                            device.apikey,
-                            newAttributes,
-                            device
-                        )
-                    );
-                }
-            }
-        }
-    }
-
-    async.series(sideEffects, callback);
-}
-
-/**
- * Extract all the commands from the attributes section and add them to the Commands Queue.
- *
- * @param {String} device           Object that contains all the information about the device.
- * @param {String} id               Entity ID of the device to find.
- * @param {String} type             Type of the device to find.
- * @param {String} service          Service of the device.
- * @param {String} subservice       Subservice of the device.
- * @param {Array}  attributes       List of attributes to update with their types and values.
- */
-function pushCommandsToQueue(device, id, type, service, subservice, attributes, callback) {
-    async.map(attributes, apply(commands.add, service, subservice, device.id), callback);
-}
-
-/**
- * Generate all the update actions corresponding to a update context request using Ngsi1.
- * Update actions include updates in attributes and execution of commands. This action will
- * be called once per Context Element in the request.
- *
- * @param {Object} req                  Update request to generate Actions from
- * @param {Object} contextElement       Context Element whose actions will be extracted.
- */
-function generateUpdateActionsNgsi1(req, contextElement, callback) {
-    function splitUpdates(device, callback) {
-        var attributes = [],
-            commands = [],
-            found;
-
-        if (device.commands) {
-            attributeLoop: for (var i in contextElement.attributes) {
-                for (var j in device.commands) {
-                    if (contextElement.attributes[i].name === device.commands[j].name) {
-                        commands.push(contextElement.attributes[i]);
-                        found = true;
-                        continue attributeLoop;
-                    }
-                }
-
-                attributes.push(contextElement.attributes[i]);
-            }
-
-        } else {
-            attributes = contextElement.attributes;
-        }
-
-        callback(null, attributes, commands, device);
-    }
-
-    function createActionsArray(attributes, commands, device, callback) {
-        var updateActions = [];
-
-        if (updateHandler) {
-            updateActions.push(
-                async.apply(
-                    updateHandler,
-                    contextElement.id,
-                    contextElement.type,
-                    req.headers['fiware-service'],
-                    req.headers['fiware-servicepath'],
-                    attributes)
-            );
-        }
-
-        if (commandHandler) {
-            if (device.polling) {
-                updateActions.push(
-                    async.apply(
-                        pushCommandsToQueue,
-                        device,
-                        contextElement.id,
-                        contextElement.type,
-                        req.headers['fiware-service'],
-                        req.headers['fiware-servicepath'],
-                        contextElement.attributes)
-                );
-            } else {
-                updateActions.push(
-                    async.apply(
-                        commandHandler,
-                        contextElement.id,
-                        contextElement.type,
-                        req.headers['fiware-service'],
-                        req.headers['fiware-servicepath'],
-                        commands)
-                );
-            }
-        }
-
-        updateActions.push(
-            async.apply(
-                executeUpdateSideEffects,
-                device,
-                contextElement.id,
-                contextElement.type,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath'],
-                contextElement.attributes)
-        );
-
-        callback(null, updateActions);
-    }
-
-    deviceService.getDeviceByName(contextElement.id, req.headers['fiware-service'], req.headers['fiware-servicepath'],
-        function(error, deviceObj) {
-            if (error) {
-                callback(error);
-            } else {
-                async.waterfall([
-                    apply(deviceService.findConfigurationGroup, deviceObj),
-                    apply(deviceService.mergeDeviceWithConfiguration,
-                        [
-                            'lazy',
-                            'internalAttributes',
-                            'active',
-                            'staticAttributes',
-                            'commands',
-                            'subscriptions'
-                        ],
-                        [null, null, [], [], [], [], []],
-                        deviceObj
-                    ),
-                    splitUpdates,
-                    createActionsArray
-                ], callback);
-            }
-        });
-}
-
-/**
- * Generate all the update actions corresponding to a update context request using Ngsi2.
- * Update actions include updates in attributes and execution of commands.
- *
- * @param {Object} req                  Update request to generate Actions from
- * @param {Object} contextElement       Context Element whose actions will be extracted.
- */
-function generateUpdateActionsNgsi2(req, contextElement, callback) {
-    var entityId;
-    var entityType;
-
-    if (contextElement.id && contextElement.type) {
-        entityId = contextElement.id;
-        entityType = contextElement.type;
-    }else if (req.params.entity) {
-        entityId = req.params.entity;
-    }
-
-    function splitUpdates(device, callback) {
-        var attributes = [],
-            commands = [],
-            found,
-            newAtt,
-            i;
-
-        if (device.commands) {
-            attributeLoop: for (i in contextElement) {
-                for (var j in device.commands) {
-
-                    if (i === device.commands[j].name) {
-                        newAtt = {};
-                        newAtt[i] = contextElement[i];
-                        newAtt[i].name = i;
-                        commands.push(newAtt[i]);
-                        found = true;
-                        continue attributeLoop;
-                    }
-                }
-            }
-
-        }
-
-        for (i in contextElement) {
-            if (i !== 'type' && i !== 'id') {
-                newAtt = {};
-                newAtt = contextElement[i];
-                newAtt.name = i;
-                attributes.push(newAtt);
-            }
-        }
-
-        callback(null, attributes, commands, device);
-    }
-
-    function createActionsArray(attributes, commands, device, callback) {
-        var updateActions = [];
-
-        if(!entityType) {
-            entityType = device.type;
-        }
-
-        if (updateHandler) {
-            updateActions.push(
-                async.apply(
-                    updateHandler,
-                    entityId,
-                    entityType,
-                    req.headers['fiware-service'],
-                    req.headers['fiware-servicepath'],
-                    attributes)
-            );
-        }
-
-        if (commandHandler) {
-            if (device.polling) {
-                updateActions.push(
-                    async.apply(
-                        pushCommandsToQueue,
-                        device,
-                        entityId,
-                        entityType,
-                        req.headers['fiware-service'],
-                        req.headers['fiware-servicepath'],
-                        attributes)
-                );
-            } else {
-                updateActions.push(
-                    async.apply(
-                        commandHandler,
-                        entityId,
-                        entityType,
-                        req.headers['fiware-service'],
-                        req.headers['fiware-servicepath'],
-                        commands)
-                );
-            }
-        }
-
-        updateActions.push(
-            async.apply(
-                executeUpdateSideEffects,
-                device,
-                entityId,
-                entityType,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath'],
-                attributes)
-        );
-
-        callback(null, updateActions);
-    }
-
-    deviceService.getDeviceByName(entityId, req.headers['fiware-service'], req.headers['fiware-servicepath'],
-        function(error, deviceObj) {
-            if (error) {
-                callback(error);
-            } else {
-                async.waterfall([
-                    apply(deviceService.findConfigurationGroup, deviceObj),
-                    apply(deviceService.mergeDeviceWithConfiguration,
-                        [
-                            'lazy',
-                            'internalAttributes',
-                            'active',
-                            'staticAttributes',
-                            'commands',
-                            'subscriptions'
-                        ],
-                        [null, null, [], [], [], [], []],
-                        deviceObj
-                    ),
-                    splitUpdates,
-                    createActionsArray
-                ], callback);
-            }
-        });
-}
-
-/**
- * Express middleware to manage incoming update context requests using NGSIv2.
- */
-function handleUpdateNgsiLD(req, res, next) {
-
-    function reduceActions(actions, callback) {
-        callback(null, _.flatten(actions));
-    }
-
-    if (updateHandler || commandHandler) {
-        logger.debug(context, 'Handling LD update from [%s]', req.get('host'));
-        logger.debug(context, req.body);
-
-        async.waterfall([
-            apply(async.map, req.body.entities, apply(generateUpdateActionsNgsi2, req)),
-            reduceActions,
-            async.series
-        ], function(error, result) {
-            if (error) {
-                logger.debug(context, 'There was an error handling the update action: %s.', error);
-
-                next(error);
-            } else {
-                logger.debug(context, 'Update action from [%s] handled successfully.', req.get('host'));
-                res.status(204).json();
-            }
-        });
-    } else {
-        logger.error(context, 'Tried to handle an update request before the update handler was stablished.');
-
-        var errorNotFound = new Error({
-            message: 'Update handler not found'
-        });
-        next(errorNotFound);
-    }
-}
-
-/**
- * Express middleware to manage incoming update context requests using NGSIv2.
- */
-function handleUpdateNgsi2(req, res, next) {
-    function reduceActions(actions, callback) {
-        callback(null, _.flatten(actions));
-    }
-
-    if (updateHandler || commandHandler) {
-        logger.debug(context, 'Handling v2 update from [%s]', req.get('host'));
-        logger.debug(context, req.body);
-
-        async.waterfall([
-            apply(async.map, req.body.entities, apply(generateUpdateActionsNgsi2, req)),
-            reduceActions,
-            async.series
-        ], function(error, result) {
-            if (error) {
-                logger.debug(context, 'There was an error handling the update action: %s.', error);
-
-                next(error);
-            } else {
-                logger.debug(context, 'Update action from [%s] handled successfully.', req.get('host'));
-                res.status(204).json();
-            }
-        });
-    } else {
-        logger.error(context, 'Tried to handle an update request before the update handler was stablished.');
-
-        var errorNotFound = new Error({
-            message: 'Update handler not found'
-        });
-        next(errorNotFound);
-    }
-}
-
-/**
- * Express middleware to manage incoming UpdateContext requests using NGSIv1.
- * As NGSI10 requests can affect multiple entities, for each one of them a call
- * to the user update handler function is made.
- */
-function handleUpdateNgsi1(req, res, next) {
-
-    function reduceActions(actions, callback) {
-        callback(null, _.flatten(actions));
-    }
-
-    if (updateHandler || commandHandler) {
-        logger.debug(context, 'Handling v1 update from [%s]', req.get('host'));
-        logger.debug(context, req.body);
-
-        async.waterfall([
-            apply(async.map, req.body.contextElements, apply(generateUpdateActionsNgsi1, req)),
-            reduceActions,
-            async.series
-        ], function(error, result) {
-            if (error) {
-                logger.debug(context, 'There was an error handling the update action: %s.', error);
-
-                next(error);
-            } else {
-                logger.debug(context, 'Update action from [%s] handled successfully.', req.get('host'));
-                res.status(200).json(createUpdateResponse(req, res, result));
-            }
-        });
-    } else {
-        logger.error(context, 'Tried to handle an update request before the update handler was stablished.');
-
-        var errorNotFound = new Error({
-            message: 'Update handler not found'
-        });
-        next(errorNotFound);
-    }
-}
-
-/**
- * Handle queries coming to the IoT Agent via de Context Provider API (as a consequence of a query to a passive
- * attribute redirected by the Context Broker).
- *
- * @param {String} id           Entity name of the selected entity in the query.
- * @param {String} type         Type of the entity.
- * @param {String} service      Service the device belongs to.
- * @param {String} subservice   Division inside the service.
- * @param {Array} attributes    List of attributes to read.
- */
-function defaultQueryHandlerNgsi1(id, type, service, subservice, attributes, callback) {
-    var contextElement = {
-        type: type,
-        isPattern: false,
-        id: id,
-        attributes: []
-    };
-
-    deviceService.getDeviceByName(id, service, subservice, function(error, ngsiDevice) {
-        if (error) {
-            callback(error);
-        } else {
-            for (var i = 0; i < attributes.length; i++) {
-                var lazyAttribute = _.findWhere(ngsiDevice.lazy, { name: attributes[i] }),
-                    command = _.findWhere(ngsiDevice.commands, { name: attributes[i] }),
-                    attributeType;
-
-                if (command) {
-                    attributeType = command.type;
-                } else if (lazyAttribute) {
-                    attributeType = lazyAttribute.type;
-                } else {
-                    attributeType = 'string';
-                }
-
-                contextElement.attributes.push({
-                    name: attributes[i],
-                    type: attributeType,
-                    value: ''
-                });
-            }
-
-            callback(null, contextElement);
-        }
-    });
-}
-
-/**
- * Handle queries coming to the IoT Agent via de Context Provider API (as a consequence of a query to a passive
- * attribute redirected by the Context Broker).
- *
- * @param {String} id           Entity name of the selected entity in the query.
- * @param {String} type         Type of the entity.
- * @param {String} service      Service the device belongs to.
- * @param {String} subservice   Division inside the service.
- * @param {Array} attributes    List of attributes to read.
- */
-function defaultQueryHandlerNgsi2(id, type, service, subservice, attributes, callback) {
-    var contextElement = {
-        type: type,
-        id: id
-    };
-
-    deviceService.getDeviceByName(id, service, subservice, function(error, ngsiDevice) {
-        if (error) {
-            callback(error);
-        } else {
-            for (var i = 0; i < attributes.length; i++) {
-                var lazyAttribute = _.findWhere(ngsiDevice.lazy, { name: attributes[i] }),
-                    command = _.findWhere(ngsiDevice.commands, { name: attributes[i] }),
-                    attributeType;
-
-                if (command) {
-                    attributeType = command.type;
-                } else if (lazyAttribute) {
-                    attributeType = lazyAttribute.type;
-                } else {
-                    attributeType = 'string';
-                }
-
-                contextElement[attributes[i]] = {
-                    type: attributeType,
-                    value: ''
-                };
-            }
-
-            callback(null, contextElement);
-        }
-    });
-}
-
-/**
- * Express middleware to manage incoming QueryContext requests using NGSIv1.
- * As NGSI10 requests can affect multiple entities, for each one of them a call
- * to the user query handler function is made.
- */
-function handleQueryNgsi1(req, res, next) {
-    function getName(element) {
-        return element.name;
-    }
-
-    function addStaticAttributes(attributes, device, contextElement, callback) {
-
-        function inAttributes(item) {
-            return item.name && attributes.indexOf(item.name) >= 0;
-        }
-
-        if (device.staticAttributes) {
-            var selectedAttributes = device.staticAttributes.filter(inAttributes);
-
-            if (selectedAttributes.length > 0) {
-                if (contextElement.attributes) {
-                    contextElement.attributes = contextElement.attributes.concat(selectedAttributes);
-                } else {
-                    contextElement.attributes = selectedAttributes;
-                }
-            }
-        }
-
-        callback(null, contextElement);
-    }
-
-    function completeAttributes(attributes, device, callback) {
-        if (attributes && attributes.length !== 0) {
-            logger.debug(context, 'Handling received set of attributes: %j', attributes);
-            callback(null, attributes);
-        } else if (device.lazy) {
-            logger.debug(context, 'Handling stored set of attributes: %j', attributes);
-            callback(null, device.lazy.map(getName));
-        } else {
-            logger.debug(context, 'Couldn\'t find any attributes. Handling with null reference');
-            callback(null, null);
-        }
-    }
-
-    function createQueryRequests(attributes, contextEntity, callback) {
-        var actualHandler;
-
-        if (queryHandler) {
-            actualHandler = queryHandler;
-        } else {
-            actualHandler = defaultQueryHandlerNgsi1;
-        }
-
-        async.waterfall([
-            apply(
-                deviceService.getDeviceByName,
-                contextEntity.id,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath']),
-            deviceService.findConfigurationGroup
-        ], function handleFindDevice(error, device) {
-            var executeCompleteAttributes = apply(
-                completeAttributes,
-                attributes,
-                device
-                ),
-                executeQueryHandler = apply(
-                    actualHandler,
-                    contextEntity.id,
-                    contextEntity.type,
-                    req.headers['fiware-service'],
-                    req.headers['fiware-servicepath']
-                ),
-                executeAddStaticAttributes = apply(
-                    addStaticAttributes,
-                    attributes,
-                    device
-                );
-
-            callback(error, apply(async.waterfall, [
-                executeCompleteAttributes,
-                executeQueryHandler,
-                executeAddStaticAttributes
-            ]));
-        });
-
-    }
-
-    function handleQueryContextRequests(error, result) {
-        if (error) {
-            logger.debug(context, 'There was an error handling the query: %s.', error);
-            next(error);
-        } else {
-            logger.debug(context, 'Query from [%s] handled successfully.', req.get('host'));
-            res.status(200).json(createQueryResponse(req, res, result));
-        }
-    }
-
-    logger.debug(context, 'Handling query from [%s]', req.get('host'));
-
-    async.waterfall([
-        apply(async.map, req.body.entities, apply(createQueryRequests, req.body.attributes)),
-        async.series
-    ], handleQueryContextRequests);
-}
-
-/**
- * Express middleware to manage incoming query context requests using NGSI-LD.
- */
-function handleQueryNgsiLD(req, res, next) {
-    function getName(element) {
-        return element.name;
-    }
-
-    function addStaticAttributes(attributes, device, contextElement, callback) {
-
-        function inAttributes(item) {
-            return item.name && attributes.indexOf(item.name) >= 0;
-        }
-
-        if (device.staticAttributes) {
-
-            var selectedAttributes = [];
-            if (attributes === undefined || attributes.length === 0) {
-                selectedAttributes = device.staticAttributes;
-            }
-            else {
-                selectedAttributes = device.staticAttributes.filter(inAttributes);
-            }
-
-            for (var att in selectedAttributes) {
-                contextElement[selectedAttributes[att].name] = {
-                    'type' : selectedAttributes[att].type,
-                    'value' : selectedAttributes[att].value
-                };
-            }
-        }
-
-        callback(null, contextElement);
-    }
-
-    function completeAttributes(attributes, device, callback) {
-        if (attributes && attributes.length !== 0) {
-            logger.debug(context, 'Handling received set of attributes: %j', attributes);
-            callback(null, attributes);
-        } else if (device.lazy) {
-            logger.debug(context, 'Handling stored set of attributes: %j', attributes);
-            var results = device.lazy.map(getName);
-            callback(null, results);
-        } else {
-            logger.debug(context, 'Couldn\'t find any attributes. Handling with null reference');
-            callback(null, null);
-        }
-    }
-
-    function finishQueryForDevice(attributes, contextEntity, actualHandler, device, callback) {
-        var contextId = contextEntity.id;
-        var contextType = contextEntity.type;
-        if(!contextId) {
-            contextId = device.id;
-        }
-
-        if(!contextType) {
-            contextType = device.type;
-        }
-
-        deviceService.findConfigurationGroup(device, function(error, group) {
-            var executeCompleteAttributes = apply(
-                completeAttributes,
-                attributes,
-                group
-                ),
-                executeQueryHandler = apply(
-                    actualHandler,
-                    contextId,
-                    contextType,
-                    req.headers['fiware-service'],
-                    req.headers['fiware-servicepath']
-                ),
-                executeAddStaticAttributes = apply(
-                    addStaticAttributes,
-                    attributes,
-                    group
-                );
-
-            async.waterfall([
-                executeCompleteAttributes,
-                executeQueryHandler,
-                executeAddStaticAttributes
-            ], callback);
-        });        
-    }
-
-    function createQueryRequest(attributes, contextEntity, callback) {
-        var actualHandler;
-        var getFunction;
-
-        if (queryHandler) {
-            actualHandler = queryHandler;
-        } else {
-            actualHandler = defaultQueryHandlerNgsi2;
-        }
-
-        if (contextEntity.id) {
-            getFunction = apply(
-                deviceService.getDeviceByName,
-                contextEntity.id,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath']);
-        } else {
-            getFunction = apply(
-                deviceService.listDevicesWithType,
-                contextEntity.type,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath'],
-                null,
-                null);            
-        }
-
-        getFunction(function handleFindDevice(error, innerDevice) {
-            let deviceList = [];
-            if (!innerDevice) {
-                return callback(new errors.DeviceNotFound(contextEntity.id));
-            }
-
-            if(innerDevice.count) {
-                if (innerDevice.count === 0) {
-                    return callback(null, []);
-                } else {
-                    deviceList = innerDevice.devices;
-                }
-            } else {
-                deviceList = [innerDevice];
-            }
-
-            async.map(deviceList, async.apply(
-                finishQueryForDevice, attributes, contextEntity, actualHandler), function (error, results) {
-                if (error) {
-                    callback(error);
-                }
-                else if(innerDevice.count) {
-                    callback(null,results);
-                } else if(Array.isArray(results) && results.length > 0){
-                    callback(null, results);
-                } else {
-                    callback(null, results);
-                }
-            });
-        });
-
-    }
-
-    function handleQueryContextRequests(error, result) {
-        if (error) {
-            logger.debug(context, 'There was an error handling the query: %s.', error);
-            next(error);
-        } else {
-            logger.debug(context, 'Query from [%s] handled successfully.', req.get('host'));
-            res.status(200).json(result);
-        }
-    }
-
-    logger.debug(context, 'Handling query from [%s]', req.get('host'));
-    var contextEntity = {};
-
-    // At the present moment, IOTA supports query request with one entity and without patterns. This is aligned
-    // with the utilization cases in combination with ContextBroker. Other cases are returned as error
-    if (req.body.entities.length !== 1)
-    {
-        logger.warn('queries with entities number different to 1 are not supported (%d found)',
-            req.body.entities.length);
-        handleQueryContextRequests({code: 400, name: 'BadRequest', message: 'more than one entity in query'});
-        return;
-    }
-    if (req.body.entities[0].idPattern)
-    {
-        logger.warn('queries with idPattern are not supported');
-        handleQueryContextRequests({code: 400, name: 'BadRequest', message: 'idPattern usage in query'});
-        return;
-    }
-
-    contextEntity.id = req.body.entities[0].id;
-    contextEntity.type = req.body.entities[0].type;
-    var queryAtts = req.body.attrs;
-    createQueryRequest(queryAtts, contextEntity, handleQueryContextRequests);
-
-}
-
-/**
- * Express middleware to manage incoming query context requests using NGSIv2.
- */
-function handleQueryNgsi2(req, res, next) {
-    function getName(element) {
-        return element.name;
-    }
-
-    function addStaticAttributes(attributes, device, contextElement, callback) {
-
-        function inAttributes(item) {
-            return item.name && attributes.indexOf(item.name) >= 0;
-        }
-
-        if (device.staticAttributes) {
-
-            var selectedAttributes = [];
-            if (attributes === undefined || attributes.length === 0) {
-                selectedAttributes = device.staticAttributes;
-            }
-            else {
-                selectedAttributes = device.staticAttributes.filter(inAttributes);
-            }
-
-            for (var att in selectedAttributes) {
-                contextElement[selectedAttributes[att].name] = {
-                    'type' : selectedAttributes[att].type,
-                    'value' : selectedAttributes[att].value
-                };
-            }
-        }
-
-        callback(null, contextElement);
-    }
-
-    function completeAttributes(attributes, device, callback) {
-        if (attributes && attributes.length !== 0) {
-            logger.debug(context, 'Handling received set of attributes: %j', attributes);
-            callback(null, attributes);
-        } else if (device.lazy) {
-            logger.debug(context, 'Handling stored set of attributes: %j', attributes);
-            var results = device.lazy.map(getName);
-            callback(null, results);
-        } else {
-            logger.debug(context, 'Couldn\'t find any attributes. Handling with null reference');
-            callback(null, null);
-        }
-    }
-
-    function finishQueryForDevice(attributes, contextEntity, actualHandler, device, callback) {
-        var contextId = contextEntity.id;
-        var contextType = contextEntity.type;
-        if(!contextId) {
-            contextId = device.id;
-        }
-
-        if(!contextType) {
-            contextType = device.type;
-        }
-
-        deviceService.findConfigurationGroup(device, function(error, group) {
-            var executeCompleteAttributes = apply(
-                completeAttributes,
-                attributes,
-                group
-                ),
-                executeQueryHandler = apply(
-                    actualHandler,
-                    contextId,
-                    contextType,
-                    req.headers['fiware-service'],
-                    req.headers['fiware-servicepath']
-                ),
-                executeAddStaticAttributes = apply(
-                    addStaticAttributes,
-                    attributes,
-                    group
-                );
-
-            async.waterfall([
-                executeCompleteAttributes,
-                executeQueryHandler,
-                executeAddStaticAttributes
-            ], callback);
-        });        
-    }
-
-    function createQueryRequest(attributes, contextEntity, callback) {
-        var actualHandler;
-        var getFunction;
-
-        if (queryHandler) {
-            actualHandler = queryHandler;
-        } else {
-            actualHandler = defaultQueryHandlerNgsi2;
-        }
-
-        if (contextEntity.id) {
-            getFunction = apply(
-                deviceService.getDeviceByName,
-                contextEntity.id,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath']);
-        } else {
-            getFunction = apply(
-                deviceService.listDevicesWithType,
-                contextEntity.type,
-                req.headers['fiware-service'],
-                req.headers['fiware-servicepath'],
-                null,
-                null);            
-        }
-
-        getFunction(function handleFindDevice(error, innerDevice) {
-            let deviceList = [];
-            if (!innerDevice) {
-                return callback(new errors.DeviceNotFound(contextEntity.id));
-            }
-
-            if(innerDevice.count) {
-                if (innerDevice.count === 0) {
-                    return callback(null, []);
-                } else {
-                    deviceList = innerDevice.devices;
-                }
-            } else {
-                deviceList = [innerDevice];
-            }
-
-            async.map(deviceList, async.apply(
-                finishQueryForDevice, attributes, contextEntity, actualHandler), function (error, results) {
-                if (error) {
-                    callback(error);
-                }
-                else if(innerDevice.count) {
-                    callback(null,results);
-                } else if(Array.isArray(results) && results.length > 0){
-                    callback(null, results);
-                } else {
-                    callback(null, results);
-                }
-            });
-        });
-
-    }
-
-    function handleQueryContextRequests(error, result) {
-        if (error) {
-            logger.debug(context, 'There was an error handling the query: %s.', error);
-            next(error);
-        } else {
-            logger.debug(context, 'Query from [%s] handled successfully.', req.get('host'));
-            res.status(200).json(result);
-        }
-    }
-
-    logger.debug(context, 'Handling query from [%s]', req.get('host'));
-    var contextEntity = {};
-
-    // At the present moment, IOTA supports query request with one entity and without patterns. This is aligned
-    // with the utilization cases in combination with ContextBroker. Other cases are returned as error
-    if (req.body.entities.length !== 1)
-    {
-        logger.warn('queries with entities number different to 1 are not supported (%d found)',
-            req.body.entities.length);
-        handleQueryContextRequests({code: 400, name: 'BadRequest', message: 'more than one entity in query'});
-        return;
-    }
-    if (req.body.entities[0].idPattern)
-    {
-        logger.warn('queries with idPattern are not supported');
-        handleQueryContextRequests({code: 400, name: 'BadRequest', message: 'idPattern usage in query'});
-        return;
-    }
-
-    contextEntity.id = req.body.entities[0].id;
-    contextEntity.type = req.body.entities[0].type;
-    var queryAtts = req.body.attrs;
-    createQueryRequest(queryAtts, contextEntity, handleQueryContextRequests);
-
-}
-
-function handleNotificationNgsi1(req, res, next) {
-
-    function checkStatus(statusCode, callback) {
-        if (statusCode.code && statusCode.code === '200') {
-            callback();
-        } else {
-            callback(new errors.NotificationError(statusCode.code));
-        }
-    }
-
-    function extractInformation(contextResponse, callback) {
-        deviceService.getDeviceByName(
-            contextResponse.contextElement.id,
-            req.headers['fiware-service'],
-            req.headers['fiware-servicepath'],
-            function(error, device) {
-                if (error) {
-                    callback(error);
-                } else {
-                    callback(null, device, contextResponse.contextElement.attributes);
-                }
-            });
-    }
-
-    function applyNotificationMiddlewares(device, values, callback) {
-        if (notificationMiddlewares.length > 0) {
-            var firstMiddleware = notificationMiddlewares.slice(0, 1)[0],
-                rest = notificationMiddlewares.slice(1),
-                startMiddleware = apply(firstMiddleware, device, values),
-                composedMiddlewares = [startMiddleware].concat(rest);
-
-            async.waterfall(composedMiddlewares, callback);
-        } else {
-            callback(null, device, values);
-        }
-    }
-
-    function createNotificationHandler(contextResponse, callback) {
-        async.waterfall([
-            apply(checkStatus, contextResponse.statusCode),
-            apply(extractInformation, contextResponse),
-            applyNotificationMiddlewares,
-            notificationHandler
-        ], callback);
-    }
-
-    function handleNotificationRequests(error) {
-        if (error) {
-            logger.error(context, 'Error found when processing notification: %j', error);
-            next(error);
-        } else {
-            res.status(200).json({});
-        }
-    }
-
-    if (notificationHandler) {
-        logger.debug(context, 'Handling notification from [%s]', req.get('host'));
-
-        async.map(req.body.contextResponses, createNotificationHandler, handleNotificationRequests);
-
-    } else {
-        var errorNotFound = new Error({
-            message: 'Notification handler not found'
-        });
-
-        logger.error(context, 'Tried to handle a notification before notification handler was established.');
-
-        next(errorNotFound);
-    }
-}
-
-function handleNotificationNgsiLD(req, res, next) {
-    function extractInformation(dataElement, callback) {
-        var atts = [];
-        for (var key in dataElement) {
-            if (dataElement.hasOwnProperty(key)) {
-                if (key !== 'id' && key !== 'type') {
-                    var att = {};
-                    att.type = dataElement[key].type;
-                    att.value = dataElement[key].value;
-                    att.name = key;
-                    atts.push(att);
-                }
-            }
-        }
-        deviceService.getDeviceByName(
-            dataElement.id,
-            req.headers['fiware-service'],
-            req.headers['fiware-servicepath'],
-            function(error, device) {
-                if (error) {
-                    callback(error);
-                } else {
-                    callback(null, device, atts);
-                }
-            });
-    }
-
-    function applyNotificationMiddlewares(device, values, callback) {
-        if (notificationMiddlewares.length > 0) {
-            var firstMiddleware = notificationMiddlewares.slice(0, 1)[0],
-                rest = notificationMiddlewares.slice(1),
-                startMiddleware = apply(firstMiddleware, device, values),
-                composedMiddlewares = [startMiddleware].concat(rest);
-
-            async.waterfall(composedMiddlewares, callback);
-        } else {
-            callback(null, device, values);
-        }
-    }
-
-    function createNotificationHandler(contextResponse, callback) {
-        async.waterfall([
-            apply(extractInformation, contextResponse),
-            applyNotificationMiddlewares,
-            notificationHandler
-        ], callback);
-    }
-
-    function handleNotificationRequests(error) {
-        if (error) {
-            logger.error(context, 'Error found when processing notification: %j', error);
-            next(error);
-        } else {
-            res.status(200).json({});
-        }
-    }
-
-    if (notificationHandler) {
-        logger.debug(context, 'Handling notification from [%s]', req.get('host'));
-        async.map(req.body.data, createNotificationHandler, handleNotificationRequests);
-
-    } else {
-        var errorNotFound = new Error({
-            message: 'Notification handler not found'
-        });
-
-        logger.error(context, 'Tried to handle a notification before notification handler was established.');
-
-        next(errorNotFound);
-    }
-
-}
-
-function handleNotificationNgsi2(req, res, next) {
-    function extractInformation(dataElement, callback) {
-        var atts = [];
-        for (var key in dataElement) {
-            if (dataElement.hasOwnProperty(key)) {
-                if (key !== 'id' && key !== 'type') {
-                    var att = {};
-                    att.type = dataElement[key].type;
-                    att.value = dataElement[key].value;
-                    att.name = key;
-                    atts.push(att);
-                }
-            }
-        }
-        deviceService.getDeviceByName(
-            dataElement.id,
-            req.headers['fiware-service'],
-            req.headers['fiware-servicepath'],
-            function(error, device) {
-                if (error) {
-                    callback(error);
-                } else {
-                    callback(null, device, atts);
-                }
-            });
-    }
-
-    function applyNotificationMiddlewares(device, values, callback) {
-        if (notificationMiddlewares.length > 0) {
-            var firstMiddleware = notificationMiddlewares.slice(0, 1)[0],
-                rest = notificationMiddlewares.slice(1),
-                startMiddleware = apply(firstMiddleware, device, values),
-                composedMiddlewares = [startMiddleware].concat(rest);
-
-            async.waterfall(composedMiddlewares, callback);
-        } else {
-            callback(null, device, values);
-        }
-    }
-
-    function createNotificationHandler(contextResponse, callback) {
-        async.waterfall([
-            apply(extractInformation, contextResponse),
-            applyNotificationMiddlewares,
-            notificationHandler
-        ], callback);
-    }
-
-    function handleNotificationRequests(error) {
-        if (error) {
-            logger.error(context, 'Error found when processing notification: %j', error);
-            next(error);
-        } else {
-            res.status(200).json({});
-        }
-    }
-
-    if (notificationHandler) {
-        logger.debug(context, 'Handling notification from [%s]', req.get('host'));
-        async.map(req.body.data, createNotificationHandler, handleNotificationRequests);
-
-    } else {
-        var errorNotFound = new Error({
-            message: 'Notification handler not found'
-        });
-
-        logger.error(context, 'Tried to handle a notification before notification handler was established.');
-
-        next(errorNotFound);
-    }
-
-}
+    contextServerUtils = require('./contextServerUtils'),
+    ngsiv1 = require('./contextServer-NGSI-v1'),
+    ngsiv2 = require('./contextServer-NGSI-v2'),
+    ngsiLD = require('./contextServer-NGSI-LD');
 
 /**
  * Sets the new user handler for Entity update requests. This handler will be called whenever an update request arrives
@@ -1335,7 +46,7 @@ function handleNotificationNgsi2(req, res, next) {
  * @param {Function} newHandler         User handler for update requests
  */
 function setUpdateHandler(newHandler) {
-    updateHandler = newHandler;
+    contextServerUtils.updateHandler = newHandler;
 }
 
 /**
@@ -1349,7 +60,7 @@ function setUpdateHandler(newHandler) {
  * @param {Function} newHandler         User handler for update requests
  */
 function setCommandHandler(newHandler) {
-    commandHandler = newHandler;
+    contextServerUtils.commandHandler = newHandler;
 }
 
 /**
@@ -1363,7 +74,7 @@ function setCommandHandler(newHandler) {
  * @param {Function} newHandler         User handler for query requests
  */
 function setQueryHandler(newHandler) {
-    queryHandler = newHandler;
+    contextServerUtils.queryHandler = newHandler;
 }
 
 /**
@@ -1377,110 +88,7 @@ function setQueryHandler(newHandler) {
  *
  */
 function setNotificationHandler(newHandler) {
-    notificationHandler = newHandler;
-}
-
-function queryErrorHandlingNgsi1(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Query NGSIv1 error [%s] handling request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        errorCode: {
-            code: code,
-            reasonPhrase: error.name,
-            details: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-        }
-    });
-}
-
-function queryErrorHandlingNgsi2(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Query NGSIv2 error [%s] handling request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        error: error.name,
-        description: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-    });
-}
-
-function queryErrorHandlingNgsiLD(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Query NGSI-LD error [%s] handling request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        error: error.name,
-        description: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-    });
-}
-
-function updateErrorHandlingNgsi1(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Update NGSIv1 error [%s] handing request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json(
-        {
-            contextResponses: [
-                {
-                    contextElement: req.body,
-                    statusCode: {
-                        code: code,
-                        reasonPhrase: error.name,
-                        details: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-                    }
-                }
-            ]
-        }
-    );
-}
-
-function updateErrorHandlingNgsi2(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Update NGSIv2 error [%s] handing request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        error: error.name,
-        description: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-    });
-}
-
-function updateErrorHandlingNgsiLD(error, req, res, next) {
-    var code = 500;
-
-    logger.debug(context, 'Update NGSI-LD error [%s] handing request: %s', error.name, error.message);
-
-    if (error.code && String(error.code).match(/^[2345]\d\d$/)) {
-        code = error.code;
-    }
-
-    res.status(code).json({
-        error: error.name,
-        description: error.message.replace(/[<>\"\'=;\(\)]/g, '')
-    });
+    contextServerUtils.notificationHandler = newHandler;
 }
 
 /**
@@ -1489,121 +97,24 @@ function updateErrorHandlingNgsiLD(error, req, res, next) {
  * @param {Object} router      Express request router object.
  */
 function loadContextRoutes(router) {
-    //TODO: remove '//' paths when the appropriate patch comes to Orion
-    var updateMiddlewaresNgsi1 = [
-            middlewares.ensureType,
-            middlewares.validateJson(updateContextTemplateNgsi1),
-            handleUpdateNgsi1,
-            updateErrorHandlingNgsi1
-        ],
-        updateMiddlewaresNgsi2 = [
-            middlewares.ensureType,
-            middlewares.validateJson(updateContextTemplateNgsi2),
-            handleUpdateNgsi2,
-            updateErrorHandlingNgsi2
-        ],
-        updateMiddlewaresNgsiLD = [
-            middlewares.ensureType,
-            middlewares.validateJson(updateContextTemplateNgsiLD),
-            handleUpdateNgsiLD,
-            updateErrorHandlingNgsiLD
-        ],
-        queryMiddlewaresNgsi1 = [
-            middlewares.ensureType,
-            middlewares.validateJson(queryContextTemplate),
-            handleQueryNgsi1,
-            queryErrorHandlingNgsi1
-        ],
-        queryMiddlewaresNgsi2 = [
-            handleQueryNgsi2,
-            queryErrorHandlingNgsi2
-        ],
-        queryMiddlewaresNgsiLD = [
-            handleQueryNgsiLD,
-            queryErrorHandlingNgsiLD
-        ],
-        updatePathsNgsi1 = [
-            '/v1/updateContext',
-            '/NGSI10/updateContext',
-            '//updateContext'
-        ],
-        updatePathsNgsi2 = [
-            '/v2/op/update',
-            '//op/update'
-        ],
-        updatePathsNgsiLD = [
-            '/ngsi-ld/v1/entities/:id/attrs/:attr'
-        ],
-        queryPathsNgsi1 = [
-            '/v1/queryContext',
-            '/NGSI10/queryContext',
-            '//queryContext'
-        ],
-        queryPathsNgsi2 = [
-            '/v2/op/query',
-            '//op/query'
-        ],
-        queryPathsNgsiLD = [
-            '/ngsi-ld/v1/entities/:id'
-        ];
-        // In a more evolved implementation, more endpoints could be added to queryPathsNgsi2
-        // according to http://fiware.github.io/specifications/ngsiv2/stable.
-
-    var i;
     if (config.checkNgsiLD()) {
-        logger.info(context, 'Loading NGSI-LD Context server routes');
-        for (i = 0; i < updatePathsNgsiLD.length; i++) {
-            router.patch(updatePathsNgsiLD[i], updateMiddlewaresNgsiLD);
-        }
-        for (i = 0; i < queryPathsNgsiLD.length; i++) {
-            router.get(queryPathsNgsiLD[i], queryMiddlewaresNgsiLD);
-        }
-        router.post('/notify', [
-            middlewares.ensureType,
-            middlewares.validateJson(notificationTemplateNgsiLD),
-            handleNotificationNgsiLD,
-            queryErrorHandlingNgsiLD
-        ]);
+        ngsiLD.loadContextRoutes(router);
     } else if (config.checkNgsi2()) {
-        logger.info(context, 'Loading NGSI-v2 Context server routes');
-        for (i = 0; i < updatePathsNgsi2.length; i++) {
-            router.post(updatePathsNgsi2[i], updateMiddlewaresNgsi2);
-        }
-        for (i = 0; i < queryPathsNgsi2.length; i++) {
-            router.post(queryPathsNgsi2[i], queryMiddlewaresNgsi2);
-        }
-        router.post('/notify', [
-            middlewares.ensureType,
-            middlewares.validateJson(notificationTemplateNgsi2),
-            handleNotificationNgsi2,
-            queryErrorHandlingNgsi2
-        ]);
+        ngsiv2.loadContextRoutes(router);
     } else {
-        logger.info(context, 'Loading NGSI-v1 Context server routes');
-        for (i = 0; i < updatePathsNgsi1.length; i++) {
-            router.post(updatePathsNgsi1[i], updateMiddlewaresNgsi1);
-        }
-        for (i = 0; i < queryPathsNgsi1.length; i++) {
-            router.post(queryPathsNgsi1[i], queryMiddlewaresNgsi1);
-        }
-        router.post('/notify', [
-            middlewares.ensureType,
-            middlewares.validateJson(notificationTemplateNgsi1),
-            handleNotificationNgsi1,
-            queryErrorHandlingNgsi1
-        ]);
+        ngsiv1.loadContextRoutes(router);
     }
 }
 
 function addNotificationMiddleware(newMiddleware) {
-    notificationMiddlewares.push(newMiddleware);
+    contextServerUtils.notificationMiddlewares.push(newMiddleware);
 }
 
 function clear(callback) {
-    notificationMiddlewares = [];
-    notificationHandler = null;
-    commandHandler = null;
-    updateHandler = null;
+    contextServerUtils.notificationMiddlewares = [];
+    contextServerUtils.notificationHandler = null;
+    contextServerUtils.commandHandler = null;
+    contextServerUtils.updateHandler = null;
 
     if (callback) {
         callback();
