+++ test/unit/ngsi-ld/examples/contextRequests/updateContextCommandError.json
@@ -1,10 +1,21 @@
-{
-    "position_info": {
-        "type": "commandResult",
-        "value": "Stalled"
-    },
+[
+  {
+    "@context": "http://context.json-ld",
     "position_status": {
-        "type": "commandStatus",
-        "value": "ERROR"
-    }
-}
+      "type": "Property",
+      "value": {
+        "@type": "commandStatus",
+        "@value": "ERROR"
+      }
+    },
+    "position_info": {
+      "type": "Property",
+      "value": {
+        "@type": "commandResult",
+        "@value": "Stalled"
+      }
+    },
+    "id": "urn:ngsi-ld:Robot:r2d2",
+    "type": "Robot"
+  }
+]
