+++ test/unit/ngsi-ld/examples/contextRequests/updateContextCommandExpired.json
@@ -1,10 +1,21 @@
-{
-    "position_info": {
-        "type": "commandResult",
-        "value": "EXPIRED"
-    },
-    "position_status": {
-        "type": "commandStatus",
-        "value": "ERROR"
-    }
-}
+[
+	{
+		"@context":"http://context.json-ld",
+		"position_status": {
+			"type":"Property",
+			"value": {
+				"@type":"commandStatus",
+				"@value":"ERROR"
+			}
+		},
+		"position_info": {
+			"type":"Property",
+			"value": {
+				"@type":"commandResult",
+				"@value":"EXPIRED"
+			}
+		},
+		"id":"urn:ngsi-ld:Robot:r2d2",
+		"type":"Robot"
+	}
+]
