+++ test/unit/ngsi-ld/examples/contextRequests/updateContextCommandFinish.json
@@ -1,10 +1,21 @@
-{
-    "position_info": {
-        "type": "commandResult",
-        "value": "[72, 368, 1]"
-    },
-    "position_status": {
-        "type": "commandStatus",
-        "value": "FINISHED"
-    }
-}
+[
+	{
+		"@context":"http://context.json-ld",
+		"position_status": {
+			"type":"Property",
+			"value": {
+				"@type":"commandStatus",
+				"@value":"FINISHED"
+			}
+		},
+		"position_info": {
+			"type":"Property",
+			"value": {
+				"@type":"commandResult",
+				"@value":"[72, 368, 1]"
+			}
+		},
+		"id":"urn:ngsi-ld:Robot:r2d2",
+		"type":"Robot"
+	}
+]
