+++ test/unit/ngsi-ld/examples/contextRequests/updateContextCommandStatus.json
@@ -1,6 +1,14 @@
-{
-    "position_status": {
-        "type": "commandStatus",
-        "value": "PENDING"
-    }
-}
+[
+	{
+		"@context":"http://context.json-ld",
+		"position_status": {
+			"type":"Property",
+			"value": {
+				"@type":"commandStatus",
+				"@value":"PENDING"
+			}
+		},
+		"id":"urn:ngsi-ld:Robot:r2d2",
+		"type":"Robot"
+	}
+]
